fix: client 페이지 분리로 빌드 오류 해결
